+++ app/components/Tag.tsx
@@ -9,7 +9,7 @@ export interface TagProps
   isActive?: boolean;
 }
 
-export default function Tag({ name, isActive, ...props }: TagProps) {
+export default function Tag({ name, isActive, className, ...props }: TagProps) {
   const activeClasses = isActive
     ? "bg-neutral-300 text-neutral-900 border-neutral-400"
     : "bg-neutral-100 text-neutral-600";
@@ -17,7 +17,7 @@ export default function Tag({ name, isActive, ...props }: TagProps) {
   return (
     <button
       type={props.type ?? "button"}
-      className={`inline-block rounded border py-0 px-2 hover:text-black ${activeClasses}`}
+      className={`inline-block rounded border py-0 px-2 hover:text-black ${activeClasses} ${className}`}
       {...props}
     >
       {name}